feat : seperate filter logic with custom function
